Projectile based weapon added
Added the base for projectile weapons:
Created script to fire the projectiles, using a pooler script to ensure better performance and organization.
Projectile is enabled at projectile spawn point and the velocity is set to add force forward based on the projectile's speed variable. If the proj hits an Entity, that is not a Player, it will disappear and do damage to the entity. Else it would just disappear.