Overhaul complete
Camera now uses the VM and more intuitive scripts in regards to camera controls.

ISSUES THAT NEED FIXING:
PlayerSM.c needs to be cleaned, as it is a patch work of fixes, new features and old code.
CAMERA JITTER WHEN ROTATING: this is a continuation of a previous jitter issue, however it is now worse.... fix!
The prefabs in scene have been disconnected, re-prefab them.